delete unused class and optimized imports
