+++ src/oop/models/requests/UserCreationRequest.java
@@ -1,9 +1,7 @@
 package oop.models.requests;
 
-import oop.SimpleBankRestApiApplication;
+import oop.models.entities.Account;
 import oop.models.entities.User;
-
-import java.time.LocalDateTime;
 
 public class UserCreationRequest {
     private String firstName;
@@ -101,5 +99,38 @@ public class UserCreationRequest {
         user.setGender(userCreationRequest.getGender());
         user.setPassword(userCreationRequest.getPassword());
         return user;
+    }
+
+    public static class AccountCreationRequest {
+        public String getUserToken() {
+            return userToken;
+        }
+
+        public void setUserToken(String userToken) {
+            this.userToken = userToken;
+        }
+
+        private String userToken;
+
+
+
+
+        public static String validate(AccountCreationRequest accountCreationRequest){
+            if(accountCreationRequest == null){
+                return "Account creation request cannot be null";
+            }
+            if(accountCreationRequest.getUserToken() == null || accountCreationRequest.getUserToken().isEmpty()){
+                return "user token cannot be null or empty";
+            }
+
+            return "Account creation request okay!";
+        }
+
+        public static Account createAccountObject(AccountCreationRequest accountCreationRequest){
+           Account account = new Account();
+
+
+            return account;
+        }
     }
 }
